backend/pyproject.toml
[project]
name = "Gradual"
version = "0.1.0"
description = "Gradual API: Remember what you did yesterday"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "apscheduler>=3.11.0",
    "authlib>=1.5.1",
    "fastapi[standard]>=0.115.12",
    "flask-cors>=4.0.0",
    "modal>=0.73.136",
    "motor>=3.7.0",
    "pdfplumber>=0.11.6",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.0",
    "python-jose[cryptography]>=3.4.0",
    "requests>=2.32.3",
]
